requirements.txt
django
djangorestframework
djangorestframework
djangorestframework-simplejwt